+++ src/main/java/ispp/project/dondesiempre/models/outfits/dto/OutfitUpdateDTO.java
@@ -1,5 +1,6 @@
 package ispp.project.dondesiempre.models.outfits.dto;
 
+import java.math.BigDecimal;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -10,7 +11,7 @@ public class OutfitUpdateDTO {
   private String description;
   private String image;
 
-  private Double discount;
+  private BigDecimal discount;
 
   private Integer index;
 }
